django-app/petesaddressbook/requirements.txt
certifi==2018.8.24
Django==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
gunicorn==19.9.0
pip==18.0
psycopg2==2.7.5
pytz==2018.5
setuptools==40.4.1
wheel==0.31.1